PROPOSAL: Storage format changes
Additions needed to metadata storage to support storage of part data in other pools/sets.

Parts have placement indices for pool and sets.

A new object type `RemoteDataType`, which is similar to `ObjectType`, but is only used to contain metadata for object part(s) that belong to other objects.

The directory name of a `RemoteDataType` should also identify that this is a remote part, so we don't read metadata for example on listing.